test(meshtastic): expand utility test coverage and reorganize suites
Move tests from `test_meshtastic_utils_coverage.py` into specialized test files and add new test cases for health probes, firmware version extraction, and coercion utilities.

- Create `test_meshtastic_utils_health.py` to handle executor lifecycles, health probes, and degraded state logic.
- Add firmware version extraction tests to `test_meshtastic_utils_metadata.py`.
- Add coercion utility tests to `test_meshtastic_utils_service.py`.
- Add edge case tests for message handling and connection checks to `test_meshtastic_utils_message_edge.py`.
- Enhance BLE and disconnect test suites with better state resetting and stale future clearing verification.
- Delete the redundant `test_meshtastic_utils_coverage.py` file.
